requirements.txt
streamlit
langchain
replicate
faiss-cpu
pypdf
accelerate
sentence_transformers
huggingface-hub
python-dotenv
streamlit_chat